+++ packages/kitchensink/hooks/index.js
@@ -1,14 +1,58 @@
 const { runHookApp } = require('@forrestjs/hooks');
 
-function ft1({ registerAction }) {
+/**
+ * This feature uses the `registerAction()` pattern and provides
+ * a custom `name` that is shown in the boot trace.
+ */
+const ft1 = ({ registerAction }) =>
+  registerAction({
+    name: 'ft1',
+    hook: '$INIT_FEATURE',
+    handler: () => console.log('ft1'),
+  });
+
+/**
+ * This feature uses the `registerAction()` pattern as well, but
+ * its name is derived by the function name's automatically
+ */
+const ft2 = ({ registerAction }) =>
   registerAction({
     hook: '$INIT_FEATURE',
-    handler: () => {},
+    handler: () => console.log('ft2'),
   });
-}
+
+/**
+ * This feature returns a single action to be registered in the form
+ * of its configuration object.
+ *
+ * The feature's name can be derived by the function name.
+ */
+const ft3 = () => ({
+  hook: '$INIT_FEATURE',
+  handler: () => console.log('ft3'),
+});
+
+/**
+ * This feature registers a single action by defining a configuration
+ * object. The name of the feature must be explicit here.
+ */
+const ft4 = {
+  name: 'ft4',
+  hook: '$INIT_FEATURE',
+  handler: () => console.log('ft4'),
+};
+
+/**
+ * This feature returns a single action defined in the array form
+ */
+const ft5 = () => ['$INIT_FEATURE', () => console.log('ft5')];
+
+/**
+ * This feature registers a single action in the array form
+ */
+const ft6 = ['$INIT_FEATURE', () => console.log('ft6'), 'ft6'];
 
 runHookApp({
   trace: 'compact',
-  services: [],
-  features: [ft1],
+  features: [ft1, ft2, ft3, ft4, ft5, ft6],
 }).catch((err) => console.error(err));