+++ packages/service-fastify/src/index.js
@@ -5,7 +5,7 @@ const initServiceHandler = require('./init-service-handler');
 const startServiceHandler = require('./start-service-handler');
 
 const service = {
-  name: hooks.SERVICE_NAME,
+  name: 'fastify',
   trace: __filename,
 };