+++ packages/hooks/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v1.21.0
 
 - Removed babel
+- Features and services names are extensively derived by their registration form and automatically prefixed